pkp/pkp-lib#10292 Refactored user interest to model and repository
