+++ classes/facades/Repo.php
@@ -44,6 +44,7 @@ use PKP\query\Repository as QueryRepository;
 use PKP\stageAssignment\Repository as StageAssignmentRepository;
 use PKP\submissionFile\Repository as SubmissionFileRepository;
 use PKP\userGroup\Repository as UserGroupRepository;
+use PKP\user\interest\Repository as UserInterestRepository;
 
 class Repo
 {
@@ -145,5 +146,10 @@ class Repo
     public static function controlledVocab(): ControlledVocabRepository
     {
         return app(ControlledVocabRepository::class);
+    }
+
+    public static function userInterest(): UserInterestRepository
+    {
+        return app(UserInterestRepository::class);
     }
 }
